Add some type hints to tricky bits
As part of making setions repeatable, a lot of the underlying mechanics
of moving answers around will change. For that reason, type hints will
be added *in moderation* where needed, to serve as in-code documentation.

This does *not* mean that we will aim for everything typed. (To type
hint, you need to import types that will only be used for annotations,
which may lead to cyclical import hell. The simplest way to avoid
cyclical imports is to put everything into a single file. That does NOT
wash. If a type needs to be imported due to annotations only, hide
it behind `if typing.TYPE_CHECKING:`, then they wont interfere at
runtime. Also, be generous with the use of
`from _future__ import annotations`.)

"mypi.ini" and ".mypi.ini" are set to be ignored by git so that the
config in "setup.cfg" can be easily overridden locally.