Improve RDA export of person data
In EasyDMP, email addresses are used as usernames. But, since it is
possible to accidentally create duplicate users, the duplicate usernames
are suffixed with a random string, making them no longer valid email
addresses. For that reason, set RDA's "mbox"-fields to email instead of
username. Also, test for RDA "name" as User.full_name as that is the happy
path. (Not all users have full_names yet so we still need a fallback.)