Change NullBooleanField to BooleanField(null=True)
NullBooleanField will be deprecated in favor of BooleanField(null=True)
in Django 4.0, and Django 3.1 complains about them. The flip works on
Django 2.2 as well so we do the flip now.

The included migration doesn't change anything in the database itself.

We're planning to squash migrations soon anyway, so it is handy to do
this now.