Mark tests that need JSONField support
Tag them with "JSONField" so they can be run separately, or excluded
explicitly, see django docs on "Testing tools":
https://docs.djangoproject.com/en/2.2/topics/testing/tools/

Alos mark them with `skipUnlessDBFeature('has_jsonb_agg')` so that they
are not run on sqlite.

When switching to Django 3.1 or newer and using the new JSONField there,
"has_jsonb_agg" should be changed to "supports_json_field".