Add API authentication by token
For CLI usage of the API, authenticating with just a token is very
convenient.

Since this is meant for CLI API usage, no attempts are made to make it
easy to get a token for an ordinary user. Tokens can be made and fetched
via the admin and via a management command and that's it.