Clean away old migrations
It turns out that if you have a replacaeble app (in our case:
easydmp_auth replaces default auth) and something that depends on
that app (in our case: django's admin), you cannot squash all the
migrations of the replaceable app. Weird things happen *in some
cases*. The review sites refuse to be migrated because the admin
migration in the migration table exists before the squashed
migration of the replaceable auth-app.

Instead of figuring out why migrating on localhost works, while
migrating in a review doesn't, I have added back in the first
migration of our auth-app and made the squash depend on that. The
alternative is probably to edit the django-migrations table and
that is much harder to do safely, not to mention having to
describe it in the UPGRADING-file. Another possibility is to fake
all migrations back to zero (resets the table) then fake them all
the way forward again.

For any future squashing of our auth app, we need to make sure to
squash from migration 2 onwards or update the initial migration
manually. Blah.