Fix bug when importing template that uses eestore
There was some unnecessary validation in the ExportSerializer that
prevented import. We need to export the primary keys of EEStoreMounts in
order to map them to the new question ids, but validators are designed
to check what is needed on implicit, automatic create and assumes the
old question ids exist unchanged. We are validating only to check that
the JSON is well-formed, not to directly write the JSON into the
database, so a validator needed to be relaxed.