Fix ordering of sections
Sections can be nested, but calculating the order if positions are
unique per super section instead of per template is expensive
(recursive!) and hard to debug. Flat, in this case, IS better than
nested.

Previously, section positions were unique per template. This commit
is the first step to get back to that, by rewriting existing orders to
be unique per template and preventing the manual setting of a position
in admin.

Instead, on creation of a new section, it is automatically given the next
available posiition, and then the order is resorterd and rewritten so
that the order is flat. If the order needs to be changed, that'll
currently have to be done with the actions "Increment position" and
"Decrement position" in the Section admin.

Things to come:
    - Improve admin UI and get rid of the actions "Increment position"
      and "Decrement position"
    - Change the unique constraint back to per template instead of per
      super section
    - Simplify what can be simplifed due to the improved unique
      constraint