Migrate: Add an identifier to each answerset
When a section can have more than one AnswerSet per Plan, the end-user
needs to be able to refer to each AnswerSet. We could use use the
primary key, but having a dedicated text-field makes it possible for the
end-user to use their own preferred identifiers, like a DOI or a nick
name.

TODO: Have a question input type that sets this identifier. This will
most likely lead to the information being stored in two locations.
However: not storing anything in AnswerSet.data would lead to ugly hacks
regarding validity and branching. Not storing it on a direct field on
the AnswerSet would lead to other ugly hacks, as in: having to check if
the section has a question with the magical input in order to find the
question id, then look up the answer. The added complexity would not be
worth it.