Migrate: Add support for nested AnswerSets
Since Sections can nest, AnswerSets must also be able to nest, *sigh*.
An example: an RDA DMP CS plan, MUST have one or more datasets, which
MAY have zero or more distributions, which MAY have zero or more
host-sections.

I don't think it is feasible to create a constraint in the db that
checks that if a section has a supersection, an answerset of that
section must 1) have a parent answerset that 2) has the child answerset
section's supersection as section. Blah.

Luckily we don't need to complicate url lookups any further since we're
already looking up the child answerset's pk, which is unique.