Make toggleable pagination version 2
V1: If page size is not set in settings or directly on class,
paging is off. Turn on by adding the page_size GET-parameter with
a value >= 1. 0 will work but paging will be off. For this to work,
page_size is hardcoded to 0 in the class. The default is off.

V2: If page size *is* set in settings or directly on class, paging
is on, but can be *turned off* by using page_size=0 as a GET-parameter.
The default is what is in settings. Also supports page_size=max, if
max_page_size is set in settings.

Regardless, a difference fron standard PageNumberPagination is that '0'
is always a valid value to page_size.